+++ generic/itclMigrate2TclCore.h
@@ -82,8 +82,8 @@ MODULE_SCOPE Tcl_Var Tcl_NewNamespaceVar(Tcl_Interp *interp, Tcl_Namespace *nsPt
 MODULE_SCOPE void Itcl_PreserveVar(Tcl_Var var);
 MODULE_SCOPE void Itcl_ReleaseVar(Tcl_Var var);
 MODULE_SCOPE int Itcl_IsCallFrameArgument(Tcl_Interp *interp, const char *name);
-MODULE_SCOPE int Itcl_GetCallVarFrameObjc(Tcl_Interp *interp);
-MODULE_SCOPE Tcl_Obj * const * Itcl_GetCallVarFrameObjv(Tcl_Interp *interp);
+MODULE_SCOPE size_t Itcl_GetCallVarFrameObjc(Tcl_Interp *interp);
+MODULE_SCOPE Tcl_Obj *const * Itcl_GetCallVarFrameObjv(Tcl_Interp *interp);
 #define Tcl_SetNamespaceResolver _Tcl_SetNamespaceResolver
 MODULE_SCOPE int _Tcl_SetNamespaceResolver(Tcl_Namespace *nsPtr,
         struct Tcl_Resolve *resolvePtr);
